Followed along in class but didn't get everything working, need to review video

+++ src/main/java/com/nscc/onlinestore/entity/Product.java
@@ -27,7 +27,7 @@ public class Product {
     private String prodStory;
 
     @DecimalMin("0.0")
-    private Double prodPrice;
+    private Long prodPrice;
 
     private String prodImgURL;
